requirements.txt
alembic==0.8.9
Authomatic==0.1.0.post1
blinker==1.4
click==6.6
facebook2==2.2.1
Flask==0.12
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
passlib==1.7.0
plaid-python==1.4.0
python-editor==1.0.3
requests==2.7.0
SQLAlchemy==1.1.4
Werkzeug==0.11.13
WTForms==2.1
